# perf_uprobe.py: Handled syntax error for Distro SLES15 SP2.