An unknown tracepoint error has occured because of syntax error.

Signed-off-by: Naveen kumar T <naveet89@in.ibm.com>