@naveen538 2020-04-30
Before patch
-------------
avocado run perf_uprobe.py
/usr/lib/python3.6/site-packages/avocado_framework-78.0-py3.6.egg/avocado/plugins/run.py:287: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast, keep-tmp and ignore-missing-references.
  FutureWarning)
Error running method "pre_tests" of plugin "fetchasset": 'Attribute' object has no attribute 'id'
JOB ID     : 611676bbcf4628619e46648aaeb51a485240a43e
JOB LOG    : /root/avocado-fvt-wrapper/results/job-2020-04-28T22.24-611676b/job.log
 (1/3) perf_uprobe.py:PerfUprobe.test_uprobe: PASS (3.78 s)
 (2/3) perf_uprobe.py:PerfUprobe.test_uprobe_return: ERROR: Command 'perf record -o /var/tmp/tmp8qajiq_x -e probe_uprobe_test:doit -aR ./uprobe_test' failed.\nstdout: b''\nstderr: b"event syntax error: 'probe_uprobe_test:doit'\n                     \\___ unknown tracepoint\n\nError:\tFile /sys/kernel/debug/tracing/e... (1.31 s)
 (3/3) perf_uprobe.py:PerfUprobe.test_uprobe_variable: PASS (2.56 s)
RESULTS    : PASS 2 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 7.76 s
JOB HTML   : /root/avocado-fvt-wrapper/results/job-2020-04-28T22.24-611676b/results.html


After patch
-----------
 avocado run perf_uprobe.py
/usr/lib/python3.6/site-packages/avocado_framework-78.0-py3.6.egg/avocado/plugins/run.py:287: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast, keep-tmp and ignore-missing-references.
  FutureWarning)
Error running method "pre_tests" of plugin "fetchasset": 'Attribute' object has no attribute 'id'
JOB ID     : 44b537cf75c713fab2f64f15db01dcf8f2864ccc
JOB LOG    : /root/avocado-fvt-wrapper/results/job-2020-04-29T19.54-44b537c/job.log
 (1/3) perf_uprobe.py:PerfUprobe.test_uprobe: PASS (3.74 s)
 (2/3) perf_uprobe.py:PerfUprobe.test_uprobe_return: PASS (2.58 s)
 (3/3) perf_uprobe.py:PerfUprobe.test_uprobe_variable: PASS (2.56 s)
RESULTS    : PASS 3 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 9.01 s
JOB HTML   : /root/avocado-fvt-wrapper/results/job-2020-04-29T19.54-44b537c/results.html